Continue work on VM

+++ src/main/java/me/xdark/recaf/jvm/instructions/InstructionHandlerIntsEquals.java
@@ -7,9 +7,9 @@ import org.objectweb.asm.tree.JumpInsnNode;
 public final class InstructionHandlerIntsEquals implements InstructionHandler<JumpInsnNode> {
 	@Override
 	public void process(JumpInsnNode instruction, ExecutionContext ctx) throws Throwable {
-		Integer v1 = ctx.popInteger();
-		Integer v = ctx.popInteger();
-		if (v.intValue() == v1.intValue()) {
+		int v1 = ctx.popInteger();
+		int v = ctx.popInteger();
+		if (v == v1) {
 			ctx.jump(instruction.label);
 		}
 	}
